# frontend/openapi.palette.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mind Weave \u2014 palettes OpenAPI subset",
    "description": "Generated by backend/scripts/dump_openapi_palette_subset.py for palette codegen.",
    "version": "0.2.0"
  },
  "paths": {
    "/api/v1/palettes/": {
      "get": {
        "tags": [
          "palettes"
        ],
        "summary": "List Palettes",
        "description": "Return all Palettes visible to the current user (user-owned + system-level).",
        "operationId": "list_palettes_api_v1_palettes__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mw_access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mw Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PalettePublic"
                  },
                  "title": "Response List Palettes Api V1 Palettes  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "palettes"
        ],
        "summary": "Create Palette",
        "description": "Create a new Palette.",
        "operationId": "create_palette_api_v1_palettes__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mw_access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mw Access Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaletteCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PalettePublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/palettes/validate": {
      "post": {
        "tags": [
          "palettes"
        ],
        "summary": "Validate Palette Import",
        "description": "Normalize palette JSON from imports: unknown keys strip with warnings; invalid CSS \u2192 422.\nPersisted palettes still require strict payloads on POST/PUT.",
        "operationId": "validate_palette_import_api_v1_palettes_validate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mw_access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mw Access Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaletteColorsValidateBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaletteValidateResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/palettes/resolve": {
      "get": {
        "tags": [
          "palettes"
        ],
        "summary": "Resolve Workflow Palette",
        "description": "Return the **effective** workflow editor palette for this user in the given context.\n\nPrecedence: ``workflow.palette_id`` (when present and visible) \u2192\n``User.settings.preferred_editor_palette_id`` \u2192 built-in **default** preset.",
        "operationId": "resolve_workflow_palette_api_v1_palettes_resolve_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workflow_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional workflow definition id; when set, its palette_id takes precedence.",
              "title": "Workflow Id"
            },
            "description": "Optional workflow definition id; when set, its palette_id takes precedence."
          },
          {
            "name": "mw_access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mw Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PalettePublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/palettes/by-slug/{slug}": {
      "get": {
        "tags": [
          "palettes"
        ],
        "summary": "Get Palette By Slug",
        "description": "Return a built-in system palette by its stable slug (e.g. default, slate).",
        "operationId": "get_palette_by_slug_api_v1_palettes_by_slug__slug__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "mw_access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mw Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PalettePublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/palettes/{id}": {
      "get": {
        "tags": [
          "palettes"
        ],
        "summary": "Get Palette",
        "description": "Return a single Palette by ID.",
        "operationId": "get_palette_api_v1_palettes__id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          },
          {
            "name": "mw_access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mw Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PalettePublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "palettes"
        ],
        "summary": "Update Palette",
        "description": "Update a user-owned Palette.",
        "operationId": "update_palette_api_v1_palettes__id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          },
          {
            "name": "mw_access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mw Access Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaletteUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PalettePublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "palettes"
        ],
        "summary": "Delete Palette",
        "description": "Delete a user-owned Palette.",
        "operationId": "delete_palette_api_v1_palettes__id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          },
          {
            "name": "mw_access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mw Access Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PalettePublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          },
          "colors": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Colors"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/WorkflowPaletteEntryOut"
            },
            "type": "array",
            "title": "Entries"
          },
          "effective_colors": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Effective Colors"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "colors",
          "created_at",
          "updated_at",
          "entries",
          "effective_colors",
          "warnings"
        ],
        "title": "PalettePublic",
        "description": "Persisted palette row plus manifest-derived projections."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PaletteCreate": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "colors": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Colors"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "PaletteCreate",
        "description": "Request body for creating a new Palette."
      },
      "PaletteColorsValidateBody": {
        "properties": {
          "colors": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Colors"
          }
        },
        "type": "object",
        "title": "PaletteColorsValidateBody",
        "description": "Body for `/palettes/validate` preview \u2014 unknown keys strip with warnings."
      },
      "PaletteValidateResult": {
        "properties": {
          "colors": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Colors"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/WorkflowPaletteEntryOut"
            },
            "type": "array",
            "title": "Entries"
          },
          "effective_colors": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Effective Colors"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "colors",
          "entries",
          "effective_colors",
          "warnings"
        ],
        "title": "PaletteValidateResult",
        "description": "Validated / normalized projection without persisting a row."
      },
      "PaletteUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "colors": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Colors"
          }
        },
        "type": "object",
        "title": "PaletteUpdate",
        "description": "Request body for updating an existing Palette (all fields optional)."
      },
      "WorkflowPaletteEntryOut": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "effective_color": {
            "type": "string",
            "title": "Effective Color"
          }
        },
        "type": "object",
        "required": [
          "key",
          "label",
          "kind",
          "effective_color"
        ],
        "title": "WorkflowPaletteEntryOut",
        "description": "One resolved workflow palette tile (manifest key + persisted coloring)."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}